% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGMMmain.r
\name{bpgmm}
\alias{bpgmm}
\alias{parsimoniousGaussianMixtureModel}
\title{Model-Based Clustering Using Baysian PGMM}
\usage{
parsimoniousGaussianMixtureModel(niter, burn, X, n, p, delta, ggamma, m,
  qVec, qnew, constraint, dVec, sVec)
}
\arguments{
\item{niter}{the number of iterations.}

\item{burn}{the number of burn in iterations.}

\item{X}{the observation matrix with size p * m.}

\item{n}{the number of observations.}

\item{delta, ggamma}{scaler hyperparameters.}

\item{m}{the number of clusters.}

\item{qVec}{the vector of the number of factors in each clusters.}

\item{qnew}{the number of factor for a new cluster.}

\item{constraint}{the pgmm constraint, a vector of length three with binary entry. For example, c(1,1,1) means the fully constraint model.}

\item{dVec}{a vector of hyperparameters with length three, shape parameters for alpha1, alpha2 and bbeta respectively.}

\item{sVec}{a vector of hyperparameters with length three, rate parameters for alpha1, alpha2 and bbeta respectively.}
}
\description{
Carries out model-based clustering using parsimonious Gaussian mixture models. MCMC are used for parameter estimation. The RJMCMC is used for model selection.
}
