% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposalPsi.R
\name{CalculateProposalPsy}
\alias{CalculateProposalPsy}
\title{CalculateProposalPsy}
\usage{
CalculateProposalPsy(hparam, thetaYList, CxyList, constraint, m, p, qVec)
}
\arguments{
\item{hparam}{hparam}

\item{thetaYList}{thetaYList}

\item{CxyList}{CxyList}

\item{constraint}{constraint}

\item{m}{the number of clusters}

\item{p}{the number of features}

\item{qVec}{the vector of the number of factors in each clusters}
}
\description{
CalculateProposalPsy
}
