% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateRatio}
\alias{calculateRatio}
\title{Log scale ratio calculation}
\usage{
calculateRatio(deno, nume)
}
\arguments{
\item{deno}{denominator.}

\item{nume}{numerator.}
}
\value{
result of ratio
}
\description{
Log scale ratio calculation
}
\examples{
deno <- log(1)
nume <- log(2)
#'
\donttest{
calculateRatio(deno, nume)
}
}
