% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateVarList}
\alias{calculateVarList}
\title{calculateVarList}
\usage{
calculateVarList(psyList, lambdaList)
}
\arguments{
\item{psyList}{list of psy}

\item{lambdaList}{list of lambda}
}
\value{
list of covariance matrix
}
\description{
calculateVarList
}
