% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{changeConstraintFormat}
\alias{changeConstraintFormat}
\title{changeConstraintFormat}
\usage{
changeConstraintFormat(strNum)
}
\arguments{
\item{strNum}{strNum}
}
\description{
changeConstraintFormat
}
\examples{
#'
\donttest{
changeConstraintFormat(c(0, 0, 0))
}
}
