% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{evaluatePriorLambda}
\alias{evaluatePriorLambda}
\title{evaluatePriorLambda}
\usage{
evaluatePriorLambda(p, m, alpha2, qVec, psy, lambda, constraint)
}
\description{
evaluate prior value for parameter Lambda
}
