% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{evaluatePriorPsi}
\alias{evaluatePriorPsi}
\title{evaluatePriorPsi}
\usage{
evaluatePriorPsi(psy, p, m, delta, bbeta, constraint)
}
\arguments{
\item{psy}{parameter}

\item{p}{the number of features}

\item{m}{the number of clusters}

\item{delta}{parameter}

\item{bbeta}{parameter}

\item{constraint}{parameter}
}
\value{
evaluated prior value for parameter Psi
}
\description{
evaluate prior value for parameter Psi
}
