% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposalLambda.R
\name{EvaluateProposalLambda}
\alias{EvaluateProposalLambda}
\title{EvaluateProposalLambda}
\usage{
EvaluateProposalLambda(
  hparam,
  thetaYList,
  CxyList,
  constraint,
  newlambda,
  m,
  qVec,
  p
)
}
\arguments{
\item{hparam}{hparam}

\item{thetaYList}{thetaYList}

\item{CxyList}{CxyList}

\item{constraint}{constraint}

\item{newlambda}{newlambda}

\item{m}{the number of clusters}

\item{qVec}{the vector of the number of factors in each clusters}

\item{p}{the number of features}
}
\description{
EvaluateProposalLambda
}
