% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposalPsi.r
\name{EvaluateProposalPsy}
\alias{EvaluateProposalPsy}
\title{EvaluateProposalPsy}
\usage{
EvaluateProposalPsy(hparam, thetaYList, CxyList, constraint, newpsy, m, p,
  qVec, delta)
}
\arguments{
\item{hparam}{hparam}

\item{thetaYList}{thetaYList}

\item{CxyList}{CxyList}

\item{constraint}{constraint}

\item{newpsy}{newpsy}

\item{m}{the number of clusters}

\item{p}{the number of features}

\item{qVec}{the vector of the number of factors in each clusters}

\item{delta}{scaler hyperparameters}
}
\description{
EvaluateProposalPsy
}
