% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{generatePriorLambda}
\alias{generatePriorLambda}
\title{generatePriorLambda}
\usage{
generatePriorLambda(p, m, alpha2, qVec, psy, constraint)
}
\arguments{
\item{p}{the number of features}

\item{m}{the number of clusters}

\item{alpha2}{hyper parameter}

\item{qVec}{parameter}

\item{psy}{parameter}

\item{constraint}{parameter}
}
\description{
evaluate prior value for parameter Lambda
}
\examples{
p <- 10
m <- 20
alpha2 <- 1.18
qVec <- rep(4, m)
delta <- 2
bbeta <- 2
constraint <- c(0, 0, 0)
psy <- generatePriorPsi(
  p,
  m,
  delta,
  bbeta,
  constraint
)
#'
\donttest{
generatePriorLambda(
  p,
  m,
  alpha2,
  qVec,
  psy,
  constraint
)
}
}
