% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getZmat}
\alias{getZmat}
\title{Tool for vector to matrix}
\usage{
getZmat(ZOneDim, m, n)
}
\arguments{
\item{ZOneDim}{a vector.}

\item{m}{the number of cluster.}

\item{n}{sample size.}
}
\value{
adjacency matrix
}
\description{
Tool for vector to matrix
}
\examples{
m <- 20
n <- 500
ZOneDim <- sample(seq_len(m), n, replace = TRUE)
#'
\donttest{
getZmat(ZOneDim, m, n)
}
}
