% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-hparam.R
\docType{class}
\name{Hparam-class}
\alias{Hparam-class}
\title{Hyperparameter set for the Bayesian PGMM sampler.}
\description{
Hyperparameter set for the Bayesian PGMM sampler.
}
\section{Slots}{

\describe{
\item{\code{alpha1}}{First Dirichlet prior parameter for component weights.}

\item{\code{alpha2}}{Second Dirichlet prior parameter for component weights.}

\item{\code{delta}}{Shape parameter used in prior updates.}

\item{\code{ggamma}}{Prior rate parameter used in covariance updates.}

\item{\code{bbeta}}{Prior scale parameter used in covariance updates.}
}}

\keyword{internal}
