% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{likelihood}
\alias{likelihood}
\title{likelihood}
\usage{
likelihood(thetaYList, ZOneDim, qqVec, muBar, X)
}
\arguments{
\item{thetaYList}{thetaYList}

\item{ZOneDim}{ZOneDim}

\item{qqVec}{qqVec}

\item{muBar}{muBar}

\item{X}{X}
}
\description{
likelihood
}
