% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listToStrVec}
\alias{listToStrVec}
\title{Convert list of string to vector of string}
\usage{
listToStrVec(stringList)
}
\arguments{
\item{stringList}{list of string}
}
\value{
vector of string
}
\description{
Convert list of string to vector of string
}
\examples{
stringList <- list("abc")
#'
\donttest{
listToStrVec(stringList)
}
}
