% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgmmStay.R
\name{stayMCMCupdate}
\alias{stayMCMCupdate}
\title{stayMCMCupdate}
\usage{
stayMCMCupdate(
  X,
  thetaYList,
  ZOneDim,
  hparam,
  qVec,
  qnew,
  dVec,
  sVec,
  constraint,
  clusInd
)
}
\arguments{
\item{X}{X}

\item{thetaYList}{thetaYList}

\item{ZOneDim}{ZOneDim}

\item{hparam}{hparam}

\item{qVec}{qVec}

\item{qnew}{qnew}

\item{dVec}{dVec}

\item{sVec}{sVec}

\item{constraint}{constraint}

\item{clusInd}{clusInd}
}
\description{
stayMCMCupdate
}
