% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sumerizeZ}
\alias{sumerizeZ}
\title{sumerizeZ}
\usage{
sumerizeZ(Zlist, index = 1:length(Zlist))
}
\arguments{
\item{Zlist}{list of Z matrix}

\item{index}{index of Zlist}
}
\value{
count mode of Z matrix
}
\description{
sumerizeZ
}
