% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-theta-y.R
\docType{class}
\name{ThetaYList}
\alias{ThetaYList}
\title{ThetaYList-class}
\description{
Parameter set for sampled PGMM component parameters.
}
\section{Slots}{

\describe{
\item{\code{tao}}{Numeric vector of component mixing weights.}

\item{\code{psy}}{List of diagonal noise covariance matrices.}

\item{\code{M}}{List of component mean vectors.}

\item{\code{lambda}}{List of component factor loading matrices.}

\item{\code{Y}}{List of latent factor score matrices.}
}}

\keyword{internal}
