% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgmmStay.R
\name{toEthetaYlist}
\alias{toEthetaYlist}
\title{Title}
\usage{
toEthetaYlist(NEthetaYList, NEZOneDim, qnew, clusInd)
}
\arguments{
\item{NEthetaYList}{NEthetaYList}

\item{NEZOneDim}{NEZOneDim}

\item{qnew}{qnew}

\item{clusInd}{clusInd}
}
\description{
Title
}
