% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePostThetaY.R
\name{updatePostThetaY}
\alias{updatePostThetaY}
\title{Update posterior theta Y list}
\usage{
updatePostThetaY(m, n, hparam, thetaYList, ZOneDim, qVec, constraint, X,
  ggamma)
}
\arguments{
\item{m}{the number of clusters.}

\item{n}{the number of observations.}

\item{hparam}{hyper parameters}

\item{thetaYList}{theta Y list}

\item{ZOneDim}{ZOneDim}

\item{qVec}{qVec}

\item{constraint}{constraint}

\item{X}{X}

\item{ggamma}{ggamma}
}
\description{
Update posterior theta Y list
}
