% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{evaluatePrior}
\alias{evaluatePrior}
\title{evaluate Prior}
\usage{
evaluatePrior(m, p, muBar, hparam, thetaYList, ZOneDim, qVec, constraint)
}
\arguments{
\item{m}{m}

\item{p}{p}

\item{muBar}{mu_bar}

\item{hparam}{hyper parameter class}

\item{thetaYList}{theta Y list}

\item{ZOneDim}{one dime of z}

\item{qVec}{q vector}

\item{constraint}{type of constraint}
}
\value{
evaluated prior value for parameter Theta and Y.
}
\description{
evaluate prior value for parameter Theta and Y.
}
\examples{
\donttest{
url <- paste0("https://github.com/lzyacht/bpgmm-examples/",
"blob/master/data/updatePostThetaY_example.RData?raw=true")
download.file(url, destfile= "updatePostThetaY_example.RData", mode = "wb")
load("updatePostThetaY_example.RData")
evaluatePrior(m, p, muBar,hparam, thetaYList, ZOneDim, qVec,constraint)
}

}
