% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{evaluatePriorLambda}
\alias{evaluatePriorLambda}
\title{evaluatePriorLambda}
\usage{
evaluatePriorLambda(p, m, alpha2, qVec, psy, lambda, constraint)
}
\arguments{
\item{p}{the number of features}

\item{m}{the number of clusters}

\item{alpha2}{hyper parameter}

\item{qVec}{the vector of the number of factors in each clusters}

\item{psy}{parameter}

\item{lambda}{parameter}

\item{constraint}{the pgmm constraint, a vector of length three with binary entry. For example, c(1,1,1) means the fully constraint model}
}
\value{
evaluated prior value for parameter Lambda
}
\description{
evaluate prior value for parameter Lambda
}
