% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{evaluatePriorPsi}
\alias{evaluatePriorPsi}
\title{evaluatePriorPsi}
\usage{
evaluatePriorPsi(psy, p, m, delta, bbeta, constraint, clusInd)
}
\arguments{
\item{psy}{parameter}

\item{p}{the number of features}

\item{m}{the number of clusters}

\item{delta}{parameter}

\item{bbeta}{parameter}

\item{constraint}{parameter}

\item{clusInd}{cluster indicator vector}
}
\description{
evaluate prior value for parameter Psi
}
\examples{
p <- 10
m <- 20
delta <- 2
bbeta <- 2
constraint <- c(0, 0, 0)
psy <- generatePriorPsi(
  p,
  m,
  delta,
  bbeta,
  constraint
)
clusInd <- rep(1, m)
#'
\donttest{
evaluatePriorPsi(
  psy,
  p,
  m,
  delta,
  bbeta,
  constraint,
  clusInd
)
}
}
