% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{generatePriorPsi}
\alias{generatePriorPsi}
\title{generatePriorPsi}
\usage{
generatePriorPsi(p, m, delta, bbeta, constraint)
}
\arguments{
\item{p}{the number of features}

\item{m}{the number of clusters}

\item{delta}{hyperparameters}

\item{bbeta}{hyperparameters}

\item{constraint}{the pgmm constraint, a vector of length three with binary entry. For example, c(1,1,1) means the fully constraint model}
}
\description{
generate prior value for parameter Psi
}
\examples{
p <- 10
m <- 20
delta <- 2
bbeta <- 2
constraint <- c(0, 0, 0)

\donttest{
generatePriorPsi(
  p,
  m,
  delta,
  bbeta,
  constraint
)
}
}
