% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperParamClass.R
\docType{class}
\name{Hparam-class}
\alias{Hparam-class}
\title{An S4 class to represent a Hyper parameter.}
\description{
An S4 class to represent a Hyper parameter.
}
\section{Slots}{

\describe{
\item{\code{alpha1}}{A numeric value}

\item{\code{alpha2}}{A numeric value}

\item{\code{delta}}{A numeric value}

\item{\code{ggamma}}{A numeric value}

\item{\code{bbeta}}{A numeric value}
}}

