% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{likelihood}
\alias{likelihood}
\title{likelihood}
\usage{
likelihood(thetaYList, ZOneDim, qqVec, muBar, X)
}
\arguments{
\item{thetaYList}{thetaYList}

\item{ZOneDim}{ZOneDim}

\item{qqVec}{qqVec}

\item{muBar}{muBar}

\item{X}{X}
}
\value{
likelihood value for parameters thetaYList
}
\description{
likelihood
}
\examples{

\donttest{
url <- paste0("https://github.com/lzyacht/bpgmm-examples/",
"blob/master/data/CalculateProposalLambda.RData?raw=true")


download.file(url, destfile= "CalculateProposalLambda.RData", mode = "wb")
load("CalculateProposalLambda.RData")
likelihood(thetaYList, ZOneDim,qVec,muBar, X)
}
}
