% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sumerizeZ}
\alias{sumerizeZ}
\title{sumerizeZ}
\usage{
sumerizeZ(Zlist, index = 1:length(Zlist))
}
\arguments{
\item{Zlist}{Zlist}

\item{index}{index}
}
\description{
sumerizeZ
}
\examples{
Zlist <- list(c(1, 2, 3), c(3, 2, 1), c(2, 2, 2))
#'
\donttest{
sumerizeZ(Zlist)
}
}
