% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summerizeRes.R
\name{summerizePgmmRJMCMC}
\alias{summerizePgmmRJMCMC}
\title{summerizePgmmRJMCMC}
\usage{
summerizePgmmRJMCMC(pgmmResList, trueCluster = NULL)
}
\arguments{
\item{pgmmResList}{result list from pgmmRJMCMC}

\item{trueCluster}{true cluster allocation}
}
\description{
summerizePgmmRJMCMC
}
