% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramClass.R
\docType{class}
\name{ThetaYList}
\alias{ThetaYList}
\title{ThetaYList-class}
\description{
Definiton of ThetaYList parameter sets
}
\section{Slots}{

\describe{
\item{\code{tao}}{A numeric vector}

\item{\code{psy}}{A list value}

\item{\code{M}}{A list value}

\item{\code{lambda}}{A list value}

\item{\code{Y}}{A list value}
}}

