% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePostThetaY.R
\name{updatePostThetaY}
\alias{updatePostThetaY}
\title{Update posterior theta Y list}
\usage{
updatePostThetaY(m, n, p, hparam, thetaYList, ZOneDim, qVec, constraint, X)
}
\arguments{
\item{m}{the number of clusters.}

\item{n}{the number of observations.}

\item{p}{the number of features.}

\item{hparam}{hyper parameters S4 class.}

\item{thetaYList}{theta Y list}

\item{ZOneDim}{ZOneDim}

\item{qVec}{qVec}

\item{constraint}{constraint}

\item{X}{X}
}
\value{
ThetaYList S4 class
}
\description{
Update posterior theta Y list
}
\examples{
\donttest{
url <- paste0("https://github.com/lzyacht/bpgmm-examples/",
"blob/master/data/updatePostThetaY_example.RData?raw=true")
download.file(url, destfile= "updatePostThetaY_example.RData", mode = "wb")
load("updatePostThetaY_example.RData")
updatePostThetaY(m, n, p, hparam, thetaYList, ZOneDim, qVec, constraint, X)
}
}
